Fix and clarify the handling of unused bits in case of a bitstring
It is perfectly fine if the trailing octet of a bitstring is 0. That
is, the ber.Base._unused_bits method should not error out in case of
8 unused bits.
Moreover, it is perfectly fine if the last octet contains more
"unused bits" (the more precise term would be "bits set to 0") than
indicated by the "unused bits" octet.