Add support of encoding and decoding set and set-of values
This is mostly a generalization of the existing code that is used
for encoding/decoding a sequence. For the purpose of testcase
specification, the encoding mapper and decoding builder also support
the utf8string type.

Note that the existing classes SequenceEncodingMapper and
SequenceDecodingBuilder are renamed in order to reflect the fact
that these classes also deal with sets (SequenceEncodingMapper ->
ContainerEncodingMapper and SequenceDecodingBuilder ->
ContainerDecodingBuilder). This is an acceptable API change.