Fix unused bits handling for a bitstring in the ber module
The old implementation of the unused bits handling is wrong: it
assumes that the lower n bits of the last octet are set to zero,
which is wrong (because X.690 does not mandate it) (where n is
the number of unused bits).
In order to fix this, remove the code that corresponds to the
wrong assumption. Moreover, adjust the bitstring API so that the
number of unused bits can be specified during the encoding and
can be retrieved during the decoding (this is a justified API
breakage).
The testcases are adjusted accordingly.