Simplify ber.Decoder._read_container
Now, the whole container "detection" logic is encapsulated in the
AbstractContainerDecodingBuilder class. The advantage is that we
can easily support tagged types, which correspond to a set or sequence
type, by providing a suitable implementation of the is_container_tag
etc. methods.
This also fixes a bug when decoding set with an indefinite length (see
the corresponding testcase).

Note: this breaks the existing API of the AbstractContainerDecodingBuilder
class (but, again, this breakage is justified).